Add exclusion functionality to get_songs()
Exclusion argument will now be used to skip over row_items that are
already present on the page. This means that gathering row-containers on
the page can be sped up if there is an existing set of songs that have
already been indexed within another function. I.e. the
download_account_history() function should be sped up if the correct set
is passed in for the exclusion argument.